requirements.txt
# Core LangGraph/LangChain dependencies - Latest compatible
langgraph==0.6.7
langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.29
langchain-openai==0.3.33
langgraph-checkpoint-postgres==2.0.23

# Databricks integration with MLflow 3.x
databricks-langchain==0.7.1
mlflow[databricks]==3.3.2

# Database and persistence
psycopg[binary,pool]==3.2.10
langchain-postgres==0.0.15

# Utilities
python-dotenv==1.0.1